organize src structure
